fix: 修正google score api分數判別

+++ RonWeb.API/Models/Shared/ReCAPTCHA.cs
@@ -4,7 +4,7 @@ namespace RonWeb.API.Models.Shared
 	public class ReCAPTCHA
 	{
 		public bool Success { get; set; } = false;
-		public int Score { get; set; } = 0;
+		public double Score { get; set; } = 0;
 		public string Action { get; set; } = string.Empty;
 		public DateTime Challenge_ts;
 		public string Hostname { get; set; } = string.Empty;
